myweb/requirements.txt
api==0.0.7
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
Django==2.2.2
django-ckeditor==6.1.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-js-asset==1.2.2
idna==2.8
inflect==2.1.0
jaraco.itertools==4.4.2
kiwisolver==1.1.0
matplotlib==3.1.1
more-itertools==7.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.25
nose==1.3.7
numpy==1.17.0
Pillow==8.2.0
protobuf==3.17.1
pygame==1.9.6
pyparsing==2.4.1.1
python-dateutil==2.8.0
python-decouple==3.4
pytz==2019.1
requests==2.21.0
secure-smtplib==0.1.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
virtualenv==16.6.0
wolframalpha==3.0.1
xmltodict==0.12.0
